+++ packages/sdk-router/src/gaszip/gasZipModule.ts
@@ -4,9 +4,8 @@ import invariant from 'tiny-invariant'
 
 import { Query, SynapseModule } from '../module'
 import { BigintIsh } from '../constants'
-import { isNativeToken } from '../utils/handleNativeToken'
+import { isNativeToken, isSameAddress } from '../utils'
 import { getGasZipQuote, getGasZipTxStatus } from './api'
-import { isSameAddress } from '../utils/addressUtils'
 
 export class GasZipModule implements SynapseModule {
   readonly address = '0x391E7C679d29bD940d63be94AD22A25d25b5A604'
