+++ packages/sdk-router/src/operations/swap.ts
@@ -9,7 +9,12 @@ import {
   applySlippageToQuery,
   applyDeadlineToQuery,
 } from '../module'
-import { handleNativeToken } from '../utils/handleNativeToken'
+import {
+  handleNativeToken,
+  TEN_MINUTES,
+  applyOptionalDeadline,
+  calculateDeadline,
+} from '../utils'
 import { SynapseSDK } from '../sdk'
 import {
   RecipientEntity,
@@ -17,11 +22,6 @@ import {
   Slippage,
   USER_SIMULATED_ADDRESS,
 } from '../swap'
-import {
-  TEN_MINUTES,
-  applyOptionalDeadline,
-  calculateDeadline,
-} from '../utils/deadlines'
 
 export type SwapQuoteV2 = {
   routerAddress: string
