+++ packages/sdk-router/src/rfq/fastBridgeRouter.ts
@@ -18,9 +18,12 @@ import {
   reduceToQuery,
 } from '../module'
 import { BigintIsh } from '../constants'
-import { getMatchingTxLog } from '../utils/logs'
-import { adjustValueIfNative } from '../utils/handleNativeToken'
-import { CACHE_TIMES, RouterCache } from '../utils/RouterCache'
+import {
+  getMatchingTxLog,
+  adjustValueIfNative,
+  CACHE_TIMES,
+  RouterCache,
+} from '../utils'
 
 // Define type alias
 export type BridgeParams = IFastBridge.BridgeParamsStruct
