+++ packages/sdk-router/src/swap/engines/paraSwapEngine.ts
@@ -5,10 +5,15 @@ import { Contract } from '@ethersproject/contracts'
 
 import erc20ABI from '../../abi/IERC20Metadata.json'
 import { IERC20Metadata as ERC20 } from '../../typechain/IERC20Metadata'
-import { getWithTimeout, postWithTimeout } from '../../utils/api'
-import { isSameAddress } from '../../utils/addressUtils'
-import { isNativeToken } from '../../utils/handleNativeToken'
-import { logExecutionTime, logger } from '../../utils/logger'
+import {
+  getWithTimeout,
+  postWithTimeout,
+  isSameAddress,
+  isNativeToken,
+  logExecutionTime,
+  logger,
+  Prettify,
+} from '../../utils'
 import { ChainProvider } from '../../router'
 import { marshallChainToken } from '../../rfq/ticker'
 import { EngineID, SlippageMax, toBasisPoints } from '../core'
@@ -21,7 +26,6 @@ import {
   SwapEngineRoute,
 } from '../models'
 import { generateAPIRoute, TransactionData } from './response'
-import { Prettify } from '../../utils/types'
 
 const PARASWAP_API_URL = 'https://api.paraswap.io'
