+++ packages/sdk-router/src/swap/engines/response.ts
@@ -1,7 +1,7 @@
 import { BigNumber } from 'ethers'
 import { AddressZero, Zero } from '@ethersproject/constants'
 
-import { isSameAddress } from '../../utils/addressUtils'
+import { isSameAddress } from '../../utils'
 import {
   AMOUNT_NOT_PRESENT,
   applySlippage,
